pyproject.toml
[project]
name = "stripe-b2-olap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-cloud-storage>=3.1.0",
    "pandas>=2.2.3",
    "pytest>=8.3.5",
]
